id 버그 수정

+++ todo-list/src/TodoList.js
@@ -1,7 +1,11 @@
-const todoObjectList = [];
-class Todo_Class {
-  constructor(item) {
-    this.ulElement = item;
+class TodoList {
+  constructor(ulSelector, buttonSelector) {
+    this.uuid = 0;
+    this.todoObjectList = [];
+    this.ulElement = document.querySelector(ulSelector);
+    this.ulElement.addEventListener('click', this.processUlEvent.bind(this));
+    this.buttonElement = document.querySelector(buttonSelector);
+    this.buttonElement.addEventListener('click', this.add.bind(this));
   }
 
   add() {
@@ -10,50 +14,67 @@ class Todo_Class {
       alert('Enter Item');
     } else {
       const todoObject = {
-        id: todoObjectList.length,
+        id: this.uuid++,
         todoText: todoInput,
         isDone: false,
       };
-      todoObjectList.unshift(todoObject);
+      this.todoObjectList.unshift(todoObject);
       this.display();
       document.querySelector('#myInput').value = '';
     }
   }
+
   deleteItem(eraseId) {
-    const elementIndex = todoObjectList.findIndex(item => item.id === eraseId);
-    todoObjectList.splice(elementIndex, 1);
+    const elementIndex = this.todoObjectList.findIndex(
+      item => item.id === eraseId
+    );
+    this.todoObjectList.splice(elementIndex, 1);
     this.display();
+  }
+
+  processUlEvent(e) {
+    const clickedElementName = e.target.nodeName;
+    const clickedElement = e.target;
+
+    if (clickedElementName === 'LI') {
+      this.toggleItem(clickedElement);
+    } else if (clickedElementName === 'BUTTON') {
+      const deleteId = clickedElement.getAttribute('data-id');
+      this.deleteItem(Number(deleteId));
+    }
+  }
+
+  toggleItem(liElement) {
+    const checkedId = liElement.getAttribute('data-id');
+    const elementIndex = this.todoObjectList.findIndex(
+      item => item.id === Number(checkedId)
+    );
+    if (this.todoObjectList[elementIndex].isDone === true) {
+      this.todoObjectList[elementIndex].isDone = false;
+      liElement.classList.remove('checked');
+    } else {
+      this.todoObjectList[elementIndex].isDone = true;
+      liElement.classList.add('checked');
+    }
   }
 
   display() {
     this.ulElement.innerHTML = '';
-    todoObjectList.forEach(item => {
+    this.todoObjectList.forEach(item => {
       const liElement = document.createElement('li');
       const delBtn = document.createElement('button');
-
       liElement.innerHTML = item.todoText;
       liElement.setAttribute('data-id', item.id);
+      this.ulElement.appendChild(liElement);
       delBtn.setAttribute('data-id', item.id);
       delBtn.classList.add('far', 'fa-trash-alt');
-      this.ulElement.appendChild(liElement);
+      if (item.isDone === true) {
+        liElement.classList.add('checked');
+      }
       liElement.appendChild(delBtn);
-
-      delBtn.addEventListener('click', e => {
-        const deleteId = e.target.getAttribute('data-id');
-        console.log(this);
-        this.deleteItem(Number(deleteId));
-      });
-      liElement.addEventListener('click', e => {
-        e.target.classList.toggle('checked');
-      });
     });
   }
 }
 
 // main
-const listSection = document.querySelector('#myUL');
-const myTodoList = new Todo_Class(listSection);
-
-document.querySelector('.addBtn').addEventListener('click', function () {
-  myTodoList.add();
-});
+const myTodoList = new TodoList('#myUL', '.addBtn');
